lab_01_03_02 replaced floats with doubles

+++ lab_01_03_02/main.c
@@ -2,11 +2,11 @@
 
 int main()
 {
-    float r1, r2, r3;
+    double r1, r2, r3;
     
-    scanf("%f%f%f", &r1, &r2, &r3);
+    scanf("%lf%lf%lf", &r1, &r2, &r3);
 
-    float resistance = (r1 * r2 * r3) / (r1 * r2 + r2 * r3 + r1 * r3);
+    double resistance = (r1 * r2 * r3) / (r1 * r2 + r2 * r3 + r1 * r3);
 
     printf("%f", resistance);
